Restructures JavaScript files to make included files more obvious
